Simplify Embed Builder error handling (#687)
Simplify the error handling of the Embed Builder and all of its
sub-builders. Instead of returning a unique error for any method that
can error, use the pattern of delayed error handling: accept potentially
invalid input and then, when building the final embed, check all of the
input and error if necessary.

Over time we've acquired feedback that the embed builder is difficult to
use. There are too many reasons for erroring, and many of the methods
can return one. By lazily returning errors when finishing the builder,
we can also perform the step of checking the total length of the content
of the embed. This removes the duplication of work.

BREAKING CHANGES: now only the `build` method of the primary builder
returns an error, with all the possibilities of the previous errors
combined.

Approved-by: BlackHoleFox <blackholefoxdev@gmail.com>
Approved-by: Cassandra McCarthy <cassie@7596ff.com>
Merged-by: Vivian Hellyer <vivian@hellyer.dev>
Signed-off-by: Vivian Hellyer <vivian@hellyer.dev>